Article (casi terminado)

+++ L10 Workspace/Acme-Newspaper/src/main/java/services/ArticleService.java
@@ -4,9 +4,13 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Date;
 
+import javax.transaction.Transactional;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 import org.springframework.util.Assert;
 import org.springframework.validation.BindingResult;
+import org.springframework.validation.Validator;
 
 import repositories.AdminRepository;
 import repositories.ArticleRepository;
@@ -16,8 +20,11 @@ import repositories.UserRepository;
 import domain.Article;
 import domain.FollowUp;
 import domain.Newspaper;
+import domain.User;
 import forms.ArticleForm;
 
+@Service
+@Transactional
 public class ArticleService {
 
 	// Managed repository
@@ -26,16 +33,22 @@ public class ArticleService {
 
 	// Supporting services
 	@Autowired
-	private NewspaperRepository newspaperRepository;
+	private NewspaperService newspaperService;
 	
 	@Autowired
-	private UserRepository userRepository;
+	private UserService userService;
 
 	@Autowired
-	private AdminRepository adminRepository;
+	private AdminService adminService;
 	
+	//TODO: relación Article-FollowUp
+//	@Autowired
+//	private FollowUpService followUpService;
+
 	@Autowired
-	private FollowUpRepository followUpRepository;
+	private Validator validator;
+	
+	
 
 	// Constructors
 	public ArticleService() {
@@ -46,7 +59,7 @@ public class ArticleService {
 	public Article create(final int newspaperId) {
 		Article res = new Article();
 		// An article is published when the corresponding newspaper is published.
-		Newspaper newspaper = this.newspaperRepository.findOne(newspaperId);
+		Newspaper newspaper = this.newspaperService.findOne(newspaperId);
 		Date publicationMoment = newspaper.getPublicationDate();
 
 		res.setPublicationMoment(publicationMoment);
@@ -57,7 +70,9 @@ public class ArticleService {
 
 	public Article save(final Article article) {
 		Assert.notNull(article);
-		// this.userRepository.checkAuthority();
+		this.checkPrincipal(article);
+		Assert.isTrue(article.getNewspaper()
+				.getPublicationDate().after(new Date(System.currentTimeMillis()-1)));
 		Article res;
 
 		res = this.articleRepository.save(article);
@@ -65,6 +80,8 @@ public class ArticleService {
 		Assert.notNull(res);
 		return res;
 	}
+
+	
 
 	public Collection<Article> findAll() {
 		Collection<Article> res = new ArrayList<Article>();
@@ -86,17 +103,17 @@ public class ArticleService {
 	}
 
 	public void delete(final Article article){
-//		this.adminRepository.checkAuthority();
+		this.adminService.checkAuthority();
 		
 		Assert.notNull(article);
 		Assert.isTrue(article.getId()!=0);
-		Assert.isTrue(this.adminRepository.exists(article.getId()));
 		
 		
 		article.getNewspaper().getArticles().remove(article);
 		article.getWriter().getArticles().remove(article);
 		for (FollowUp followUp : article.getFollowUps()){
-			this.followUpRepository.delete(followUp);
+			//TODO: borrar followUp de Article
+//			this.followUpRepository.delete(followUp);
 		}
 		
 		this.articleRepository.delete(article);
@@ -104,7 +121,14 @@ public class ArticleService {
 	
 	// Other busines methods
 	
+	private void checkPrincipal(Article article) {
+		User principal = this.userService.findByPrincipal();
+		Assert.isTrue(principal.equals(article.getNewspaper().getPublisher()));
+		
+	}
+	
 	public ArticleForm construct(Article article){
+		Assert.notNull(article);
 		ArticleForm res = new ArticleForm();
 		
 		res.setId(article.getId());
@@ -112,7 +136,7 @@ public class ArticleService {
 		res.setSummary(article.getSummary());
 		res.setBody(article.getBody());
 		res.setPictures(article.getPictures());
-		res.setFinal(article.getIsFinal());
+		res.setIsFinal(article.getIsFinal());
 		res.setNewspaperId(article.getNewspaper().getId());
 		
 		return res;
@@ -120,17 +144,24 @@ public class ArticleService {
 	
 	public Article reconstruct(ArticleForm articleForm, BindingResult binding){
 		Assert.notNull(articleForm);
-		Newspaper newspaper = this.newspaperRepository.findOne(articleForm.getNewspaperId());
+		Newspaper newspaper = this.newspaperService.findOne(articleForm.getNewspaperId());
+		Article res;
 		
-		Article res = create(newspaper.getId());
+		if(articleForm.getId()!=0)
+			res = this.findOne(articleForm.getId());
+		else	
+			res = this.create(newspaper.getId());
 		
 		res.setId(articleForm.getId());
 		res.setTitle(articleForm.getTitle());
 		res.setSummary(articleForm.getSummary());
 		res.setBody(articleForm.getBody());
 		res.setPictures(articleForm.getPictures());
-		res.setIsFinal(articleForm.isFinal());
+		res.setIsFinal(articleForm.getIsFinal());
 		res.setNewspaper(newspaper);
+		
+		if(binding!=null)
+			this.validator.validate(res,binding);
 		
 		return res;
 	}
